requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
celery-redbeat==2.0.0
flower==0.9.7
humanize==3.2.0
importlib-metadata==3.4.0
kombu==4.6.11
prometheus-client==0.8.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
six==1.15.0
tenacity==6.3.1
tornado==6.1
typing-extensions==3.7.4.3
vine==1.3.0
zipp==3.4.0
